Renaming packages from co.cask to io.cdap
- Also updated dependencies to Spark 2.x
